requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
cycler==0.10.0
decorator==4.4.2
face-alignment==1.3.4
ffmpeg-python==0.2.0
future==0.18.2
gTTS==2.2.3
idna==3.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.4.2
networkx==2.5.1
nltk==3.6.2
numba==0.53.1
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
pkg_resources==0.0.0
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.2
pyttsx3==2.90
PyWavelets==1.1.1
regex==2021.7.6
requests==2.26.0
scikit-image==0.18.2
scikit-video==1.1.11
scipy==1.7.0
sda==0.2
six==1.16.0
tifffile==2021.7.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
